refactor: extract png writer module
